schema fixes

+++ backend/migrations/order_details.sql
@@ -2,7 +2,7 @@ CREATE TABLE orderDetails (
   orderId INT,
   productName TEXT NOT NULL,
   unitCode TEXT,
-  quantity INT NOT NULL,
-  unitPrice INT NOT NULL,
+  quantity FLOAT NOT NULL,
+  unitPrice FLOAT NOT NULL,
   FOREIGN KEY (orderId) REFERENCES orders(id)
 );
